requirements.txt
qwen-vl-utils==0.0.10
transformers-stream-generator==0.0.4
torchvision==0.19.0
git+https://github.com/huggingface/transformers.git
accelerate
av
